style(server): creates isOnOrganizerList method a whole chuk of connecting to google cloud and checking for organizer is done in a new method
